Create custom-wordpress-avatar
As a web developer, I often find myself working with WordPress. While it is a powerful platform, there are always areas where it could be improved. One lacking feature is the ability to upload and add your own avatar to your profile. Unfortunately, WordPress defaults to Gravatar for this feature. However, utilizing this simple piece of code in my functions.php file and the Advanced Custom Fields plugin, I am able to create a custom solution for users to add a custom image to their website profile. This allows for greater flexibility and customization for users, and it helps to make their website more personal. I’m excited to share this solution with other developers who may be looking for a similar resolution.