Create password-protect-page
WordPress is a great content management system because it is so versatile. You can use it for a simple blog or a complex website with many different pages and features. One of the great things about WordPress is that you can easily protect any page on your site with a password. This can be useful if you want to limit access to certain pages to specific users. For example, you might want to password protect a page that contains sensitive information. To do this, simply add the following code to your functions.php file.

This will add password protection to any page on your WordPress site and will redirect users to log in to get access. This is a great way to keep your site secure and ensure that only authorized users can access certain pages.